+++ aggligator/src/agg/task.rs
@@ -2,7 +2,9 @@
 
 use atomic_refcell::AtomicRefCell;
 use bytes::Bytes;
-use futures::{Future, Sink, Stream, StreamExt, future, stream, stream::FuturesUnordered};
+use futures::{
+    Future, FutureExt, Sink, Stream, StreamExt, future, future::BoxFuture, stream, stream::FuturesUnordered,
+};
 use rand::{prelude::*, rngs::SmallRng};
 use std::{
     collections::{HashSet, VecDeque},
@@ -20,10 +22,7 @@ use tokio::{
     select,
     sync::{mpsc, oneshot, watch},
 };
-use wokio::{
-    task::{BoxFuture, MaybeSend, MaybeSendFutureExt, MaybeSync},
-    time::{Instant, interval_stream, sleep_until, timeout},
-};
+use wokio::time::{Instant, interval_stream, sleep_until, timeout};
 
 use crate::{
     agg::link_int::{DisconnectInitiator, LinkInt, LinkIntEvent, LinkTest},
@@ -233,13 +232,7 @@ enum SendTerminate {
 }
 
 /// Link filter function type.
-trait LinkFilterCb<TAG>: FnMut(Link<TAG>, Vec<Link<TAG>>) -> BoxFuture<'static, bool> + MaybeSend {}
-impl<T, TAG> LinkFilterCb<TAG> for T where
-    T: FnMut(Link<TAG>, Vec<Link<TAG>>) -> BoxFuture<'static, bool> + MaybeSend
-{
-}
-
-type LinkFilterFn<TAG> = Box<dyn LinkFilterCb<TAG>>;
+type LinkFilterFn<TAG> = Box<dyn FnMut(Link<TAG>, Vec<Link<TAG>>) -> BoxFuture<'static, bool> + Send>;
 
 /// Task managing a connection of aggregated links.
 ///
@@ -356,9 +349,9 @@ impl<TX, RX, TAG> fmt::Debug for Task<TX, RX, TAG> {
 
 impl<TX, RX, TAG> Task<TX, RX, TAG>
 where
-    RX: Stream<Item = Result<Bytes, io::Error>> + Unpin + MaybeSend + 'static,
-    TX: Sink<Bytes, Error = io::Error> + Unpin + MaybeSend + 'static,
-    TAG: fmt::Display + MaybeSend + MaybeSync + 'static,
+    RX: Stream<Item = Result<Bytes, io::Error>> + Unpin + Send + 'static,
+    TX: Sink<Bytes, Error = io::Error> + Unpin + Send + 'static,
+    TAG: fmt::Display + Send + Sync + 'static,
 {
     #[allow(clippy::too_many_arguments)]
     pub(crate) fn new(
@@ -412,7 +405,7 @@ where
             established: None,
             stats_tx,
             stats_last_sent: Instant::now(),
-            link_filter: Box::new(|_, _| async { true }.maybe_boxed()),
+            link_filter: Box::new(|_, _| async { true }.boxed()),
             init_links: links.into(),
             refused_links_tasks: FuturesUnordered::new(),
             fatal_connect_error_rx,
@@ -670,7 +663,7 @@ where
                         .iter_mut()
                         .enumerate()
                         .filter_map(|(id, link_opt)| {
-                            link_opt.as_mut().map(|link| async move { (id, link.event().await) }.maybe_boxed())
+                            link_opt.as_mut().map(|link| async move { (id, link.event().await) }.boxed())
                         })
                         .collect();
                     tasks.shuffle(&mut fast_rng);
@@ -781,7 +774,7 @@ where
                                     .await;
                                     link.notify_disconnected(DisconnectReason::LinkFilter);
                                 }
-                                .maybe_boxed(),
+                                .boxed(),
                             );
                         } else {
                             link.notify_disconnected(DisconnectReason::LinkFilter);
@@ -2266,10 +2259,10 @@ where
     /// blocked. It should thus execute quickly.
     pub fn set_link_filter<F, Fut>(&mut self, mut link_filter: F)
     where
-        F: FnMut(Link<TAG>, Vec<Link<TAG>>) -> Fut + MaybeSend + 'static,
-        Fut: Future<Output = bool> + MaybeSend + 'static,
+        F: FnMut(Link<TAG>, Vec<Link<TAG>>) -> Fut + Send + 'static,
+        Fut: Future<Output = bool> + Send + 'static,
     {
-        self.link_filter = Box::new(move |link, others| link_filter(link, others).maybe_boxed());
+        self.link_filter = Box::new(move |link, others| link_filter(link, others).boxed());
     }
 
     /// Enables dumping of analysis data over the provided channel while the aggregator task is running.
@@ -2326,16 +2319,16 @@ where
 
 impl<TX, RX, TAG> IntoFuture for Task<TX, RX, TAG>
 where
-    RX: Stream<Item = Result<Bytes, io::Error>> + Unpin + MaybeSend + MaybeSync + 'static,
-    TX: Sink<Bytes, Error = io::Error> + Unpin + MaybeSend + MaybeSync + 'static,
-    TAG: fmt::Display + MaybeSend + MaybeSync + 'static,
+    RX: Stream<Item = Result<Bytes, io::Error>> + Unpin + Send + Sync + 'static,
+    TX: Sink<Bytes, Error = io::Error> + Unpin + Send + Sync + 'static,
+    TAG: fmt::Display + Send + Sync + 'static,
 {
     type Output = Result<(), TaskError>;
 
     type IntoFuture = BoxFuture<'static, Result<(), TaskError>>;
 
     fn into_future(self) -> Self::IntoFuture {
-        self.run().maybe_boxed()
+        self.run().boxed()
     }
 }