+++ aggligator/src/control.rs
@@ -17,10 +17,7 @@ use tokio::{
     io::{AsyncRead, AsyncWrite},
     sync::{Mutex, mpsc, watch},
 };
-use wokio::{
-    task::{MaybeSend, MaybeSync},
-    time::{Instant, error::Elapsed, timeout},
-};
+use wokio::time::{Instant, error::Elapsed, timeout};
 use x25519_dalek::{PublicKey, StaticSecret};
 
 use crate::{
@@ -313,7 +310,7 @@ impl<TX, RX, TAG> Control<TX, RX, TAG>
 where
     RX: Stream<Item = Result<Bytes, io::Error>> + Unpin,
     TX: Sink<Bytes, Error = io::Error> + Unpin,
-    TAG: fmt::Display + MaybeSend + MaybeSync + 'static,
+    TAG: fmt::Display + Send + Sync + 'static,
 {
     /// Adds a new outgoing, packet-based link to the connection.
     ///
@@ -440,7 +437,7 @@ impl<R, W, TAG> Control<IoTx<W>, IoRx<R>, TAG>
 where
     R: AsyncRead + Unpin,
     W: AsyncWrite + Unpin,
-    TAG: fmt::Display + MaybeSend + MaybeSync + 'static,
+    TAG: fmt::Display + Send + Sync + 'static,
 {
     /// Adds a new outgoing, stream-based link to the connection.
     ///