+++ aggligator/src/transport/connector.rs
@@ -1,7 +1,11 @@
 //! Link connector.
 
 use async_trait::async_trait;
-use futures::{FutureExt, StreamExt, future, stream::FuturesUnordered};
+use futures::{
+    FutureExt, StreamExt,
+    future::{self, BoxFuture},
+    stream::FuturesUnordered,
+};
 use std::{
     collections::HashSet,
     fmt,
@@ -13,10 +17,7 @@ use std::{
 };
 use tokio::sync::{RwLock, broadcast, mpsc, oneshot, watch};
 use tracing::Instrument;
-use wokio::{
-    task::{BoxFuture, MaybeSend, MaybeSendFutureExt, MaybeSync},
-    time::sleep,
-};
+use wokio::time::sleep;
 
 use super::{BoxControl, BoxLink, BoxLinkError, BoxTask, LinkCfgFn, LinkTag, LinkTagBox};
 use crate::{
@@ -26,9 +27,8 @@ use crate::{
 };
 
 /// A transport for connecting to remote endpoints.
-#[cfg_attr(not(all(target_family = "wasm", feature = "js")), async_trait)]
-#[cfg_attr(all(target_family = "wasm", feature = "js"), async_trait(?Send))]
-pub trait ConnectingTransport: MaybeSend + MaybeSync + 'static {
+#[async_trait]
+pub trait ConnectingTransport: Send + Sync + 'static {
     /// Name of the transport.
     fn name(&self) -> &str;
 
@@ -65,9 +65,8 @@ pub trait ConnectingTransport: MaybeSend + MaybeSync + 'static {
 type ArcConnectingTransport = Arc<dyn ConnectingTransport>;
 
 /// A wrapper for an outgoing link.
-#[cfg_attr(not(all(target_family = "wasm", feature = "js")), async_trait)]
-#[cfg_attr(all(target_family = "wasm", feature = "js"), async_trait(?Send))]
-pub trait ConnectingWrapper: MaybeSend + MaybeSync + fmt::Debug + 'static {
+#[async_trait]
+pub trait ConnectingWrapper: Send + Sync + fmt::Debug + 'static {
     /// Name of the wrapper.
     fn name(&self) -> &str;
 
@@ -133,9 +132,7 @@ impl ConnectorBuilder {
     ///
     /// This is called each time before a link is connected and may modify
     /// the link-specific configuration.
-    pub fn set_link_cfg(
-        &mut self, link_cfg_fn: impl Fn(&dyn LinkTag, &mut LinkCfg) + MaybeSend + MaybeSync + 'static,
-    ) {
+    pub fn set_link_cfg(&mut self, link_cfg_fn: impl Fn(&dyn LinkTag, &mut LinkCfg) + Send + Sync + 'static) {
         self.link_cfg_fn = Arc::new(link_cfg_fn);
     }
 
@@ -321,8 +318,8 @@ impl Connector {
             let tags_changed = future::select_all(
                 transport_tags
                     .iter_mut()
-                    .map(|tt| tt.changed().maybe_boxed())
-                    .chain(iter::once(future::pending().maybe_boxed())),
+                    .map(|tt| tt.changed().boxed())
+                    .chain(iter::once(future::pending().boxed())),
             );
 
             enum ConnectorEvent {
@@ -587,6 +584,6 @@ impl IntoFuture for ConnectingTransportHandle {
                 Err(_) => Ok(()),
             }
         }
-        .maybe_boxed()
+        .boxed()
     }
 }