+++ aggligator-transport-websocket-web/CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to the Aggligator WebSocket transport for the web will be do
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Changed
+- the WebSocket is wrapped into a thread-bound proxy again, since aggligator
+  requires transports to be `Send` and `Sync` on the web again
+- the configuration function passed to `WebSocketConnector::set_cfg` needs to be
+  `Send` and `Sync` again
+
+### Added
+- dependency on threadporter
+
 ## 0.6.0 - 2026-08-13
 ### Changed
 - the configuration function passed to `WebSocketConnector::set_cfg` no longer